file path

+++ www/explore/include/lib/CLEAN.php
@@ -15,14 +15,14 @@ class CLEAN
 	{
 		$files = scandir(dirname(__FILE__) . "/../../");
 
-		echo "<pre>"; var_dump(dirname(__FILE__) . "/../../"); echo "</pre>";
-		echo "<pre>"; var_dump($files); echo "</pre>";
+		//echo "<pre>"; var_dump(dirname(__FILE__) . "/../../"); echo "</pre>";
+		//echo "<pre>"; var_dump($files); echo "</pre>";
 
 		foreach ($files as $key => $filename) 
 		{
 			$filetype = explode(".", $filename);
 
-			if ($filetype[1] == "json") 
+			if (isset($filetype[1]) && $filetype[1] == "json") 
 			{
 				unlink($filename);
 			}
